# api.yaml
openapi: 3.1.0
info:
  title: 库存管理系统接口
  version: 0.0.1
servers:
  - url: 'https://development.gigantic-server.com/v1'
    description: Development server
  - url: 'https://staging.gigantic-server.com/v1'
    description: Staging server
  - url: 'https://api.gigantic-server.com/v1'
    description: Production server
security:
  - oauth2: []
  - bearerAuth: []
paths:
...
  /inventory/inbound:
    post:
      summary: 创建入库单
      operationId: createInbound
      tags:
        - 库存
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Inbound'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inbound'
  '/inventory/inbound/{id}':
    get:
      summary: 入库单详情
      operationId: getInbound
      tags:
        - 库存
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
            description: 入库单ID
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inbound'
    put:
      summary: 更新入库单
      operationId: updateInbound
      tags:
        - 库存
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
            description: 入库单ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Inbound'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inbound'
    delete:
      summary: 删除入库单
      operationId: deleteInbound
      tags:
        - 库存
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
            description: 入库单ID
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
  '/inventory/inbound/{id}/approve':
    post:
      summary: 确认入库单
      operationId: approveInbound
      tags:
        - 库存
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
            description: 入库单ID
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inbound'
  '/inventory/inbound/{id}/reject':
    post:
      summary: 拒绝入库单
      operationId: rejectInbound
      tags:
        - 库存
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
            description: 入库单ID
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inbound'
...
  /inventory/inboundReport/generate:
    post:
      summary: 生成入库报告
      operationId: generateInboundReport
      tags:
        - 入库报告
      description: |
        生成指定入库记录的入库报告PDF/Excel。
        报告将异步生成，返回报告ID用于查询生成状态和下载。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InboundReportReq'
      responses:
        '202':
          description: 入库报告生成任务已接受
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InboundReportResp'
        '400':
          description: 请求参数错误
        '404':
          description: 入库记录不存在
        '500':
          description: 服务器内部错误
  '/inventory/inboundReports/{reportId}/status':
    get:
      summary: 获取入库报告生成状态
      operationId: getInboundReportStatus
      tags:
        - 入库报告
      description: |
        根据报告ID获取入库报告的生成状态。
        可用于轮询检查报告是否生成完成。
      parameters:
        - name: reportId
          in: path
          required: true
          description: 入库报告ID
          schema:
            type: string
      responses:
        '200':
          description: 成功获取入库报告状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InboundReport'
        '404':
          description: 报告不存在
        '500':
          description: 服务器内部错误
  '/inventory/inboundReports/{reportId}/download':
    get:
      summary: 下载入库报告
      operationId: downloadInboundReport
      tags:
        - 入库报告
      description: |
        下载已生成的入库报告文件。
        支持PDF和Excel格式的报告下载。
      parameters:
        - name: reportId
          in: path
          required: true
          description: 入库报告ID
          schema:
            type: string
      responses:
        '200':
          description: 成功下载入库报告文件
          content:
            application/pdf:
              schema:
                type: string
                format: binary
              examples:
                pdf_report:
                  summary: PDF格式入库报告
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
              examples:
                excel_report:
                  summary: Excel格式入库报告
          headers:
            Content-Disposition:
              description: 文件下载附件头
              schema:
                type: string
                example: attachment; filename="inbound_report_20240115.pdf"
            Content-Length:
              description: 文件大小
              schema:
                type: integer
        '404':
          description: 报告不存在或尚未生成完成
        '410':
          description: 下载链接已过期
        '500':
          description: 服务器内部错误
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        password:
          tokenUrl: /auth/login
          scopes: {}
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StorageFeePlan:
      type: object
      properties:
        id:
          type: integer
          description: 保管费用计划ID
        fromDate:
          type: string
          description: 开始日期
        payerName:
          type: string
          description: 付款方名称
    Inbound:
      type: object
      properties:
        id:
          type: integer
          description: 入库记录ID
        inboundOrderId:
          type: string
          description: 入库订单ID
        inboundDate:
          type: string
          format: date
          description: 入库日期
        customerOrderId:
          type: string
          description: 客户订单号
        warehouse:
          type: object
          properties:
            id:
              type: integer
              description: 仓库ID
            name:
              type: string
              description: 仓库名称
        customer:
          type: object
          properties:
            id:
              type: integer
              description: 客户ID
            name:
              type: string
              description: 客户名称
        items:
          type: array
          description: 入库物品列表
          items:
            $ref: '#/components/schemas/InboundItem'
        storageFeePlans:
          type: array
          description: 保管费用计划列表
          items:
            $ref: '#/components/schemas/StorageFeePlan'
        status:
          type: string
          description: 入库状态
          enum:
            - draft
            - pending
            - approved
            - rejected
            - cancelled
        createdAt:
          type: string
          description: 创建时间
...
    InboundItem:
      type: object
      properties:
        id:
          type: integer
          description: 入库物品ID
        inboundId:
          type: integer
          description: 入库记录ID
        inbound:
          type: object
          properties:
            id:
              type: integer
              description: 入库记录ID
            inboundOrderId:
              type: string
              description: 入库订单ID
            inboundDate:
              type: string
              description: 入库日期
            warehouse:
              type: object
              properties:
                id:
                  type: integer
                  description: 仓库ID
                name:
                  type: string
                  description: 仓库名称
        productId:
          type: integer
          description: 商品ID
        product:
          type: object
          properties:
            id:
              type: integer
              description: 商品ID
            name:
              type: string
              description: 商品名称
        quantity:
          type: integer
          description: 入库数量
        perItemWeight:
          type: number
          description: 单品重量
        perItemWeightUnit:
          type: string
          description: 单品重量单位
        totalWeight:
          type: number
          description: 总重量
        manufactureDate:
          type: string
          format: date
          description: 生产日期
        bestBeforeDate:
          type: string
          format: date
          description: 最佳风味期限
        lotNumber:
          type: string
          description: 批次号
        shipName:
          type: string
          description: 船名
...
    InboundReportReq:
      type: object
      properties:
        inboundId:
          type: integer
          description: 入库记录ID
        format:
          type: string
          description: 报告格式
          enum:
            - pdf
            - excel
          default: pdf
      required:
        - inboundId
    InboundReportResp:
      type: object
      properties:
        reportId:
          type: string
          description: 入库报告ID
        reportDate:
          type: string
          format: date-time
          description: 报告生成时间
        warehouse:
          type: object
          properties:
            id:
              type: integer
              description: 仓库ID
            name:
              type: string
              description: 仓库名称
        customer:
          type: object
          properties:
            id:
              type: integer
              description: 客户ID
            name:
              type: string
              description: 客户名称
        expiresAt:
          type: string
          format: date-time
          description: 下载链接过期时间
    InboundReport:
      type: object
      properties:
        id:
          type: string
          description: 入库报告ID
        warehouse:
          type: object
          properties:
            id:
              type: integer
              description: 仓库ID
            name:
              type: string
              description: 仓库名称
        customer:
          type: object
          properties:
            id:
              type: integer
              description: 客户ID
            name:
              type: string
              description: 客户名称
        reportDate:
          type: string
          format: date
          description: 报告日期
        format:
          type: string
          description: 报告格式
          enum:
            - pdf
            - excel
        status:
          type: string
          description: 报告状态
          enum:
            - pending
            - processing
            - completed
            - failed
        downloadUrl:
          type: string
          description: 报告下载URL（仅在completed状态时提供）
        errorMessage:
          type: string
          description: 错误信息（仅在failed状态时提供）
        createdAt:
          type: string
          format: date-time
          description: 报告创建时间
        completedAt:
          type: string
          format: date-time
          description: 报告完成时间
        expiresAt:
          type: string
          format: date-time
          description: 下载链接过期时间
...